fix: typo and move to cart feature
