+++ frontend/src/context/cartContext.js
@@ -11,13 +11,11 @@ const CartProvider = ({ children }) => {
   const localCartItems = localStorage.getItem("cartItems")
     ? JSON.parse(localStorage.getItem("cartItems"))
     : [];
-
   // use reducer hooks
   const [state, dispatch] = useReducer(cartReducer, {
     cartItems: localCartItems,
   });
   const[total, setTotal] = useState(0)
-
 
   // setting items to the localStorage
   useEffect(() => {
@@ -44,9 +42,8 @@ const CartProvider = ({ children }) => {
     removeItemFromCart(id, dispatch)
     toast.success("Item Removed")
   }
-
   return (
-    <cartContext.Provider value={{ state, addToCart, removeFromCart , total}}>
+    <cartContext.Provider value={{ state, addToCart, removeFromCart,  total}}>
       {children}
     </cartContext.Provider>
   );
