+++ frontend/src/pages/cart/cart.css
@@ -3,7 +3,7 @@
     border: none;
     color: var(--love-dark);
     padding: 5px;
-    font-size: 1rem;
+    font-size: 0.8rem;
     margin-top: 1rem;
     transition: transform .2s
 }
